// src/app/models/arbol.model.ts
export interface Arbol {
    id_arbol: number;
    nombre_comun: string;
    nombre_cientifico: string;
    altura: number;
    diametro_tronco: number;
    edad_estimada: number;
    estado: string;
    tipo_hoja: string;
    especie: string;
    uso_ecologico: string;
    oxigeno_anual: number;
    epoca_floracion: string;
    notas: string;
    latitud: number;
    longitud: number;
    area: string;
    imagen_url: string;
  }
  